Creacion de archivo DateUtils
Contiene las funciones necesarias para mostrar una fecha según su precisión.
